Rename utils.py to xml_utils.py
utils.py file contains only XmlUtils class functionality so be in
consistance with other utils* files it was renamed to xml_utils.py
