+++ metadata.py
@@ -1,4 +1,4 @@
-import utils
+from xml_utils import XmlUtils
 
 import metadata_versioning
 
@@ -20,13 +20,13 @@ class Metadata:
 
     @staticmethod
     def parse(input_string):
-        doc = utils.XmlUtils.parse(input_string)
+        doc = XmlUtils.parse(input_string)
         self = Metadata()
-        self.__groupId      = utils.XmlUtils.get_value_by_path(doc.documentElement, "/groupId")
-        self.__artifactId   = utils.XmlUtils.get_value_by_path(doc.documentElement, "/artifactId")
-        self.__version      = utils.XmlUtils.get_value_by_path(doc.documentElement, "/version")
+        self.__groupId      = XmlUtils.get_value_by_path(doc.documentElement, "/groupId")
+        self.__artifactId   = XmlUtils.get_value_by_path(doc.documentElement, "/artifactId")
+        self.__version      = XmlUtils.get_value_by_path(doc.documentElement, "/version")
 
-        versioning          = utils.XmlUtils.get_item_by_path(doc.documentElement, "/versioning")
+        versioning          = XmlUtils.get_item_by_path(doc.documentElement, "/versioning")
         self.__versioning   = metadata_versioning.MetadataVersioning.from_element(versioning)
 
         return self
